rename dir Views views
