--- requirements.txtY
@@ -1,5 +1,0 @@
-^C
-^C
-xX
-^X
-
